+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.3
+
+- search: add market parameter
+- add request rate handling with retries
+
 ## 0.3.2
 
 - users: move `me` calls to own endpoint class